Bugfix: Reservation and RoomView
Reservation view not saving when its initiated from RoomView fixed.

+++ src/business/ReservationManager.java
@@ -6,14 +6,25 @@ import entity.Reservation;
 import java.util.List;
 
 public class ReservationManager {
-    private ReservationDAO reservationDAO;
+    private final ReservationDAO reservationDAO;
+    private final RoomManager roomManager;
 
     public ReservationManager() {
         this.reservationDAO = new ReservationDAO();
+        this.roomManager = new RoomManager();
+    }
+
+    public List<Reservation> getAllReservationsWithHotelName() {
+        return reservationDAO.getAllReservationsWithHotelName();
+    }
+
+    public Reservation getReservationById(int id) {
+        return reservationDAO.getReservationById(id);
     }
 
     public void addReservation(Reservation reservation) {
         reservationDAO.addReservation(reservation);
+        roomManager.decreaseRoomStock(reservation.getRoomId(), 1);
     }
 
     public void updateReservation(Reservation reservation) {
@@ -21,14 +32,10 @@ public class ReservationManager {
     }
 
     public void deleteReservation(int id) {
-        reservationDAO.deleteReservation(id);
-    }
-
-    public Reservation getReservationById(int id) {
-        return reservationDAO.getReservationById(id);
-    }
-
-    public List<Reservation> getAllReservationsWithHotelName() {
-        return reservationDAO.getAllReservationsWithHotelName();
+        Reservation reservation = reservationDAO.getReservationById(id);
+        if (reservation != null) {
+            reservationDAO.deleteReservation(id);
+            roomManager.increaseRoomStock(reservation.getRoomId(), 1);
+        }
     }
 }
